kafka-producer-image/requirements.txt
faker
kafka-python
pandas